requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
ccxt>=4.1.0
pydantic>=2.5.0
pandas>=2.1.0
numpy>=1.25.0
jinja2>=3.1.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
python-multipart>=0.0.6
sse-starlette>=1.6.5
aiohttp>=3.8.0
colorama>=0.4.4
